Correct copyright year in license Correct 'actor' occurrences in comments
