+++ bindings/swift/xcode/IotaWallet/IotaWallet.xcodeproj/xcshareddata/xcschemes/IotaWallet.xcscheme
@@ -10,7 +10,7 @@
             ActionType = "Xcode.IDEStandardExecutionActionsCore.ExecutionActionType.ShellScriptAction">
             <ActionContent
                title = "Run Script"
-               scriptText = "# Type a script or drag a script file from your workspace to insert its path.&#10;cd ${PROJECT_DIR}&#10;src=../../target/debug/libiota_wallet.dylib&#10;dest=iota_wallet/libiota_wallet.dylib&#10;dest_dir=iota_wallet&#10;cp &quot;../../iota_wallet_ffi.h&quot; &quot;$dest_dir&quot;&#10;if [ &quot;$src&quot; -nt &quot;$dest&quot; ]; then&#10;  rm ${PROJECT_DIR}/prebuild.log &#10;  cp &quot;$src&quot; &quot;$dest_dir&quot;&#10;else&#10;  exec &gt; ${PROJECT_DIR}/prebuild.log 2&gt;&amp;1&#10;  cargo build &amp;&amp; cp &quot;$src&quot; iota_wallet&#10;fi&#10;"
+               scriptText = "# Type a script or drag a script file from your workspace to insert its path.&#10;cd ${PROJECT_DIR}&#10;src=../../target/debug/libiota_wallet.dylib&#10;dest=iota_wallet/libiota_wallet.dylib&#10;dest_dir=iota_wallet&#10;cp &quot;../../iota_wallet_ffi.h&quot; &quot;$dest_dir&quot;&#10;if [ &quot;$src&quot; -nt &quot;$dest&quot; ]; then&#10;  rm ${PROJECT_DIR}/prebuild.log &#10;  cp &quot;$src&quot; &quot;$dest_dir&quot;&#10;else&#10;  exec &gt; ${PROJECT_DIR}/prebuild.log 2&gt;&amp;1&#10;  cargo build &amp;&amp; cp &quot;$src&quot; &quot;$dest_dir&quot;&#10;fi&#10;"
                shellToInvoke = "/bin/zsh">
                <EnvironmentBuildable>
                   <BuildableReference
