+++ bindings/swift/xcode/IotaWallet/IotaWallet/Wallet.h
@@ -1,3 +1,5 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
 
 #import <Foundation/Foundation.h>
