+++ bindings/swift/xcode/IotaWallet/IotaWalletTests/IotaWalletTests-Bridging-Header.h
@@ -1,4 +1,6 @@
-//
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
