+++ bindings/swift/xcode/IotaWallet/iota_wallet/iota_wallet_ffi.h
@@ -1,3 +1,6 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
 #ifdef __cplusplus
 extern "C" {
 #endif
