+++ src/lib.rs
@@ -10,7 +10,7 @@
 pub mod account;
 /// The account manager module.
 pub mod account_manager;
-/// The actor interface for the library. A different way to call the wallet functions, useful for bindings to other
+/// The message passing interface for the library. A different way to call the wallet functions, useful for bindings to other
 /// languages.
 #[cfg(feature = "message_interface")]
 pub mod message_interface;
