+++ src/message_interface/account_method.rs
@@ -1,4 +1,4 @@
-// Copyright 2020 IOTA Stiftung
+// Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::account::operations::{address_generation::AddressGenerationOptions, syncing::SyncOptions};
