+++ src/message_interface/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2020 IOTA Stiftung
+// Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 mod account_method;
@@ -33,4 +33,25 @@ pub async fn send_message(handle: &WalletMessageHandler, message_type: MessageTy
     let message = Message::new(message_type, message_tx);
     handle.handle(message).await;
     message_rx.recv().await.unwrap()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{AccountToCreate, MessageType, ResponseType};
+
+    #[tokio::test]
+    async fn create_account() {
+        let wallet_handle = super::create_message_handler(None).await.unwrap();
+
+        // create an account
+        let account = AccountToCreate { alias: None };
+        let response = super::send_message(&wallet_handle, MessageType::CreateAccount(Box::new(account))).await;
+        match response.response() {
+            ResponseType::CreatedAccount(account) => {
+                let id = account.id().clone();
+                println!("Created account id: {id}")
+            }
+            _ => panic!("unexpected response {:?}", response),
+        }
+    }
 }
